+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. [19246](https://github.com/influxdata/influxdb/pull/19246): Redesign load data page to increase discovery and ease of use
 1. [19334](https://github.com/influxdata/influxdb/pull/19334): Add --active-config flag to influx to set config for single command
 1. [19219](https://github.com/influxdata/influxdb/pull/19219): List buckets via the API now supports after (ID) parameter as an alternative to offset.
+1. [19390](https://github.com/influxdata/influxdb/pull/19390): Record last success and failure run times in the Task
 
 ### Bug Fixes